+++ src/hooks/useUser.ts
@@ -6,33 +6,34 @@ import { useLoginContext } from '@/context/LoginProvider';
 type User = Tables<'users'>;
 
 export const useUser = () => {
+  const { login } = useLoginContext();
   const [user, setUser] = useState<User | null>(null);
-  const [loading, setLoading] = useState(true);
-  const { isLogin } = useLoginContext();
+  const [loading, setLoading] = useState<boolean>(true);
+
+  const supabase = createClient();
+
+  const getUser = async () => {
+    try {
+      const {
+        data: { session }
+      } = await supabase.auth.getSession();
+
+      if (session?.user) {
+        const { data } = await supabase.from('users').select('*').eq('id', session.user.id).single();
+        setUser(data);
+        login();
+      } else {
+        setUser(null);
+      }
+    } catch (err) {
+      console.error('Error fetching user:', err);
+      setUser(null);
+    } finally {
+      setLoading(false);
+    }
+  };
 
   useEffect(() => {
-    const supabase = createClient();
-
-    const getUser = async () => {
-      try {
-        const {
-          data: { session }
-        } = await supabase.auth.getSession();
-
-        if (session?.user) {
-          const { data } = await supabase.from('users').select('*').eq('id', session.user.id).single();
-          setUser(data);
-        } else {
-          setUser(null);
-        }
-      } catch (err) {
-        console.error(err);
-        setUser(null);
-      } finally {
-        setLoading(false);
-      }
-    };
-
     const {
       data: { subscription }
     } = supabase.auth.onAuthStateChange((event, session) => {
@@ -49,7 +50,7 @@ export const useUser = () => {
     return () => {
       subscription.unsubscribe();
     };
-  }, [isLogin]);
+  }, [login]);
 
   return { user, loading };
 };
